Always drop count for '}' line
It's not of much use in general.  It's clear that destructors are
called at the end of the scope, but for coverage we can live without
this peace of data.  Moreover if there is some global static object
with public destructor, its count will be assigned to last '}' line,
which makes it look very weird.

This reduces number of relevant lines, so there should be more
useful coverage information.