Generate and validate hash of source files
To avoid calculating them in the application we calculate them as
git does for its blobs and compare results directly with object ids.

This is an optimization, which might go away.  Calculating the hash
ourselves is not hard, so can remove this to avoid this a bit strange
dependency on git's implementation detail.